Extraigo lectura del data frame fuera de DistributionIndexer
